Optimise ssh and ws utils
