+++ pkg/util/ssh/ssh.go
@@ -1,0 +1,42 @@
+/*
+Copyright 2021 The Pixiu Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package ssh
+
+import (
+	"fmt"
+	"time"
+
+	"golang.org/x/crypto/ssh"
+
+	"github.com/caoyingjunz/pixiu/pkg/types"
+)
+
+func NewSSHClient(sshConfig *types.WebSSHRequest) (*ssh.Client, error) {
+	// TODO：利用 gin 的解析，直接设置默认值
+	port := sshConfig.Port
+	if port == 0 {
+		port = 22
+	}
+
+	// TODO 补充支持 PrivateKey 场景
+	return ssh.Dial("tcp", fmt.Sprintf("%s:%d", sshConfig.Host, port), &ssh.ClientConfig{
+		Timeout:         time.Second * 5,
+		User:            sshConfig.User,
+		Auth:            []ssh.AuthMethod{ssh.Password(sshConfig.Password)},
+		HostKeyCallback: ssh.InsecureIgnoreHostKey(), // 忽略 know_hosts 检查
+	})
+}
